+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * [Added] `nodata` option (bool, string, or callback)
   * [Added] ext-table `mergeStatusColumns` option
   * [Added] new method `tree.enable(flag)`
+  * [Added] new method `tree.expandAll(flag, opts)`
   * [Fixed] ES6 import dependency on jquery for jquery.fancytree.ui-deps.js
   * [Fixed] #863 setActive() sometimes does not scroll node into view
   * [Fixed] #877 postProcess may now also return the object form `{..., children: []}`